refactor: create prisma folder and move prisma.schema into it; create deploy command
