+++ app/database/index.ts
@@ -1,10 +1,10 @@
-import { PrismaClient } from "@prisma/client";
+import { PrismaClient } from "prisma/generated/client";
 import { Observer } from "./observer";
 
 const client = new PrismaClient();
 export const observer = new Observer(client);
-
 export const prisma = observer.prisma;
-export const {
-  // hello: Hello
+
+export const { 
+  user: User
 } = prisma;
