+++ prisma/schema.prisma
@@ -5,4 +5,5 @@ datasource db {
 
 generator client {
   provider = "prisma-client-js"
+  output   = "generated/client"
 }
